# OpenMP: No memset in viewfill