--- a/core/src/Kokkos_CopyViews.hpp
+++ b/core/src/Kokkos_CopyViews.hpp
@@ -1350,7 +1350,8 @@ contiguous_fill_or_memset(
     typename ViewTraits<DT, DP...>::const_value_type& value) {
 // On A64FX memset seems to do the wrong thing with regards to first touch
 // leading to the significant performance issues
-#ifndef KOKKOS_ARCH_A64FX
+#if !defined(KOKKOS_ARCH_A64FX) && !defined(KOKKOS_ENABLE_OPENMP) && \
+    !defined(KOKKOS_ENABLE_SERIAL)
   if (Impl::is_zero_byte(value))
     ZeroMemset<ExecutionSpace, View<DT, DP...>>(exec_space, dst, value);
   else