@masterleinad 2023-11-02
I'm at least surprised that a manual loop would be faster than `memset` for the `Serial` backend.