@rgayatri23 2023-11-02
> I'm at least surprised that a manual loop would be faster than `memset` for the `Serial` backend.

Sorry I was wrong on that one, its the opposite, `memset` is 2x faster for `Serial` backend. I corrected it.