@fnrizzi 2023-11-09
maybe it makes sense to add this to the benchmarks?