@masterleinad 2023-11-09
We should probably do the same for the `View` construction, see `Kokkos_ViewMapping.hpp` (but that can be in a follow-up).