@masterleinad 2023-11-09
Seeing
```
n = 100000, Fill test -1
Write time: 0.02s; Write Bandwidth: 0.05GB/s
Write time: 0.00s; Write Bandwidth: 100.29GB/s
Write time: 0.00s; Write Bandwidth: 120.08GB/s
Write time: 0.00s; Write Bandwidth: 112.74GB/s
Write time: 0.00s; Write Bandwidth: 130.25GB/s
n = 100000, Fill test 0
Write time: 0.00s; Write Bandwidth: 2.09GB/s
Write time: 0.00s; Write Bandwidth: 36.28GB/s
Write time: 0.00s; Write Bandwidth: 48.19GB/s
Write time: 0.00s; Write Bandwidth: 49.36GB/s
Write time: 0.00s; Write Bandwidth: 49.61GB/s
n = 1000000, Fill test -1
Write time: 0.00s; Write Bandwidth: 2.73GB/s
Write time: 0.01s; Write Bandwidth: 1.00GB/s
Write time: 0.00s; Write Bandwidth: 2.02GB/s
Write time: 0.00s; Write Bandwidth: 777.30GB/s
Write time: 0.00s; Write Bandwidth: 711.74GB/s
n = 1000000, Fill test 0
Write time: 0.00s; Write Bandwidth: 3.83GB/s
Write time: 0.00s; Write Bandwidth: 16.67GB/s
Write time: 0.00s; Write Bandwidth: 25.60GB/s
Write time: 0.00s; Write Bandwidth: 25.60GB/s
Write time: 0.00s; Write Bandwidth: 25.02GB/s
n = 10000000, Fill test -1
Write time: 0.04s; Write Bandwidth: 2.21GB/s
Write time: 0.01s; Write Bandwidth: 6.77GB/s
Write time: 0.00s; Write Bandwidth: 85.59GB/s
Write time: 0.00s; Write Bandwidth: 169.72GB/s
Write time: 0.00s; Write Bandwidth: 351.79GB/s
n = 10000000, Fill test 0
Write time: 0.01s; Write Bandwidth: 7.08GB/s
Write time: 0.01s; Write Bandwidth: 14.39GB/s
Write time: 0.01s; Write Bandwidth: 14.39GB/s
Write time: 0.01s; Write Bandwidth: 14.24GB/s
Write time: 0.01s; Write Bandwidth: 14.39GB/s
```
on `SapphireRapids` (104 cores).